+++ .travis.yml
@@ -1,0 +1,8 @@
+language: python
+python:
+  - "2.5"
+  - "2.6"
+  - "2.7"
+
+install: pip install 'GitPython >= 0.2.0' --use-mirrors
+script: nosetests test/unit
